fix: address CodeQL security findings from the first Go scan (27 findings) (#425)
Retargeting PR #424 to go-main and fixing codeql.yml (language: python -> go)
surfaced the first-ever CodeQL analysis of the Go port. 27 findings, addressed here:

Fixed (real code changes):
- go/zipslip, go/unsafe-unzip-symlink (3, goldenreplay/archive.go): the tar
  extraction I wrote for the testdata-compression work didn't validate entry
  paths or symlink targets against escaping destDir. Added safeJoin (a
  standard zip-slip guard) plus rejection of absolute/escaping symlink
  targets.
- go/path-injection (3, source_map.go): source-map lookups built a filesystem
  path from a JS stack-trace URL (externally-submitted RUM/error telemetry)
  without validating it stayed within SOBS_SOURCE_MAP_DIR. Applied the same
  "prefix with / before Clean" guard already used by handleStatic for
  /static/.
- go/incorrect-integer-conversion (13 of 14): otlp_ingest.go, handlers_v1_ingest.go,
  handlers_pages_logs_errors.go, fix_rum_helpers.go, fix_mcp_helpers.go all
  narrowed an int64 parsed from untrusted OTLP/RUM/MCP input to int before any
  bounds check. Fixed by keeping values at int64 width (jsonenc renders int
  and int64 identically, so output bytes are unaffected) or, where a bound
  exists, clamping at int64 width before the final narrowing (new clampInt64
  helper). dbutil.go's cInt (1 finding) is suppressed instead — ~120 call
  sites across the handler package make widening it a much larger, riskier
  change than the finding's practical severity justifies (every deployment
  target is amd64/arm64, where int is 64-bit; strconv.ParseInt already
  clamps rather than wraps on overflow).

Suppressed with inline // codeql[...] justification comments (verified, not
just asserted):
- go/weak-sensitive-data-hashing (2, query_exec.go, handlers_pages_logs_errors.go):
  MD5 here is a content-addressing fingerprint for trace/record IDs embedded
  directly in golden-corpus response bytes, mirroring app.py's own
  hashlib.md5() use for the same IDs — changing the algorithm would break
  byte-parity, and MD5 isn't protecting secret/sensitive data in this
  non-cryptographic-ID role.
- go/reflected-xss (5, handlers_forms.go x2, handlers_pages.go, handlers_misc.go,
  server.go): every flagged output already goes through this port's own
  HTML-escaping — htmlEscapeMarkup (handlers_forms.go's redirect confirmation
  pages) or render.Engine's Jinja-compatible autoescape (internal/render/escape.go,
  mirroring MarkupSafe.escape() for every {{ }} template value). CodeQL's Go
  query doesn't recognize either as a sanitizer, so it flags the eventual
  w.Write as an unescaped sink despite the escaping already having happened.

Verified: go build/vet/test clean; full golden-corpus suite still passes
(Docker, CI-pinned libchdb v26.5.0) — none of these changes affect output
bytes for any in-range input.

+++ go/cmd/sobs/coverage_pure_a_test.go
@@ -208,7 +208,7 @@ func TestExtractTraceFields(t *testing.T) {
 		event     map[string]any
 		wantTrace string
 		wantSpan  string
-		wantFlags int
+		wantFlags int64
 	}{
 		{
 			"empty event",
